softusb: clear EPs on disconnect
Along with the logic to handle multiple interfaces, commit
a26dc5116da806b5180958b0d5007d318c2a887b also introduced the following
bug:

When a keyboard or composite device with keyboard was replaced by a
mouse after enumeration, the stack would still poll the "keyboard",
which most likely resulted in the mouse data to be b drained and
discarded.

This patch clears the EP roles on disconnect.