libfpvm: make scanner aware of function names and handle arity in parser
This patch makes the scanner recognize function names as distinct tokens
(and not just as identifiers). This gives the parser more information
and will also enable further changes.

The parser now only accepts functions with the correct number of
arguments. Before, mistakes like sin(a, b) or min(a) were silently
accepted.